Medspace/Advance__Tumor_Diagnosis_hospfinder_X-ray-main/BrainTumor Classification DL/static/images/healthcare_chatbot (final)2/healthcare_chatbot/requirements.txt
flask
pymongo
python-dotenv
groq
PyPDF2
werkzeug
